io: let system to define invalid logical address
Invalid logical address may be different between systems.
E.g. 0 is not a valid virtual address from Linux userspace.
However, for baremetal system, logical address is the physical
address.

In the common metal_io_virt() function which is used to get the virtual
address, remove checking if the logical address is valid if no
invalid logical address is defined.

Signed-off-by: Wendy Liang <jliang@xilinx.com>