CMake: lib/include/ instead of lib/ as include dir
Using the lib/ as include directory when compiling the libmetal
can cause issues if the header files names conflict to those
in the system. And thus, need to use a different directory as the
include directory such that the headers in the lib/ will not be
included.

The headers of libmetal library needs to be included as "metal/xxxx.h"

Signed-off-by: Wendy Liang <jliang@xilinx.com>