system:freertos:zynqmp_r5: correct io mem mapping
This fix is to workaround the memory mapping in for ZynqMP_R5.

The root cause is in the libxil.a MPU region setting.
The base address for the region setitng in libxil is
calculated as this:

region_base = addr & ~(region_size - 1);
This will cause issue if the input addr address is
in the end of the region.

Signed-off-by: Wendy Liang <jliang@xilinx.com>