io: remove io_mem_map user API
We should have I/O memory map API for user as
the memory should already be mapped when the I/O region
is ready to use.

However, as I/O memory mapping is still required on RTOS
and baremetal systems, we keep the I/O memory mapping
for library internal usage.

Signed-off-by: Wendy Liang <jliang@xilinx.com>