mutex: cleanup how we init mutexes
There are assumptions that one could directly assign a value to init a
mutex which shouldn't be made.  Instead we should use the
metal_mutex_init call to setup the mutex.  Also for static or global
defines of a mutex we need a way to set them up.  So we introduce
METAL_MUTEX_DEFINE(m) for that case.

Signed-off-by: Kumar Gala <kumar.gala@linaro.org>