lib: zephyr: log: provide a zephyr specific METAL_INIT_DEFAULTS
Define a Zephyr specific METAL_INIT_DEFAULTS to set log handler to
metal_zephyr_log_handler by default.  Now that we set that in the
defaults, we also remove overriding the handler if the user explicitly
passes in NULL.  This way the user can still have a NULL handler if
they don't want any logging.

Signed-off-by: Kumar Gala <kumar.gala@linaro.org>