mutex: Provide a macro to init a static structure member that is a mutex
Provide a means to initialize a structure element that is a mutex that
happens to be global or static.  This can save us from having to do the
init at runtime (and thus not optimize away some code if the struct
isn't used at all).

So we now have METAL_MUTEX_DEFINE to define & init a global list named
foobar:

METAL_MUTEX_DEFINE(foobar);

and METAL_MUTEX_INIT to do:

struct foo {
	int a;
	metal_mutex_t x;
};

struct foo bar = {
	.a = 4,
	.x = METAL_MUTEX_INIT(bar.x),
};

Signed-off-by: Kumar Gala <kumar.gala@linaro.org>
Signed-off-by: Wendy Liang <jliang@xilinx.com>