Move libmetal Wiki link to Overview section
This commit fixes a duplicated definite article 'the' in the last
section and moves the corrected paragraph to the "Overview" section.
This change is made because the paragraph contains the Wiki link that
provides valuable information about libmetal internals, including a
detailed section titled "OpenAMP Overview".

It also addresses multiple format issues, including missing or duplicate
spaces in the text, and corrects the line length to match the format
used in README.md.

Signed-off-by: Gustavo Romero <gustavo.romero@linaro.org>